+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/String+Localized.swift
@@ -198,4 +198,8 @@ extension String.Localized {
         STPLocalizedString("Confirm the payment in your bank or payment app within %@ to complete the purchase.",
                            "Text for alert message when user needs to confirm payment in their banking app")
     }
+    
+    static var cpf_cpnj: String {
+        STPLocalizedString("CPF/CPNJ", "Label for CPF/CPNJ (Brazil tax ID) field")
+    }
 }